Update fuzzing_directed.h
The code was confused between a bound pointer to a member (i.e., the address of this member in this instance) and a pointer-to-member (i.e., the offset of this member within the class). A recent build of the Visual C++ Compiler (correctly) complains about the attempt to create the pointer-to-member as the member in question is protected - it also complains as it cannot bring the two sides of the ':' to a common type.